Rendering to multiple windows is supported
Currently graphics context and state is not shared between windows. This could be optimized, for example sharing GL context (textures, geometry etc), environment maps, and shadow maps between windows.

+++ src/Skybolt/SkyboltEngine/VisHud.cpp
@@ -175,7 +175,7 @@ VisHud::~VisHud()
 {
 }
 
-void VisHud::setAspectRatio(double value)
+void VisHud::setAspectRatio(float value)
 {
 	mCamera->setProjectionMatrix(osg::Matrix::ortho2D(-value, value, -1, 1));
 }
